requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
Flask==3.0.0
Flask-Login==0.6.3
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
pandas==2.2.0
plotly==5.18.0
xlsxwriter==3.1.9
gunicorn==21.2.0